Update Python unit tests
Added a bash script to create a dummy, test Git repository. From now on,
this will be used in the Python unit tests.

I also added one extra member method in GitRepository and a few new
tests. It's just easier to test things with the test repo and I've added
this extra change as a proof-of-concept.